operatorBtn, equalBtn, clearBtn
added a clearBtn for when AC is clicked to wipe
textcontent on display and stored value in num1,
num2, and mathOperator. a check function to skip
the equalBtn operation if num1, num2, or operator
hasn't been chosen yet. Otherwise operate with the
stored information. if operatorBtn is pressed when
all the needed information is already there then
operate with existing info and store the result in
num1 and replace the operator with the new info.